\name{B20}
\alias{B20}
\title{Farnsworth B-20 cap colors}
\usage{B20}
\description{\code{B20} contains the cap colors for the Farnsworth B-20 test, in XYZ coordinates.
The Farnsworth B-20 is a short test for detecting congenital color vision deficiencies.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{CapNo}{ Cap Number}
    \item{Munsell}{ Munsell color}
    \item{X}{ CIE X cap color}
    \item{Y}{ CIE Y cap color}
    \item{Z}{ CIE Z cap color}
 }
}
\source{
Judd, D.B. and MacAdam, D.L., 1979
Contributions to Color Science
University of Rochester. Institute of Optics and Center for Building Technology
Department of Commerce, National Bureau of Standards
}
\references{
Judd, D.B. and MacAdam, D.L., 1979
Contributions to Color Science
University of Rochester. Institute of Optics and Center for Building Technology
Department of Commerce, National Bureau of Standards
}
\author{Jose Gama}
\examples{
data(B20)
B20
}
\keyword{datasets}
