\name{dichromaticCopunctalPoint}
\alias{dichromaticCopunctalPoint}
\title{Copunctal points derived by Smith and Pokorny (1975)}
\usage{dichromaticCopunctalPoint}
\description{ \code{dichromaticCopunctalPoint} contains the copunctal points derived by Smith and Pokorny (1975)}
\format{
  This data frame contains the following columns:
  \describe{
    \item{P}{ copunctal points - protanope}
    \item{D}{ copunctal points - deuteranope}
    \item{T}{ copunctal points - tritanope}
  }
}
\source{
Smith, V. C. & Pokorny, J.
Spectral sensitivity of the foveal cone photopigments between
400 and 500 nm. Vision Research, 15, 1975. 161-171.
}
\references{
Smith, V. C. & Pokorny, J.
Spectral sensitivity of the foveal cone photopigments between
400 and 500 nm. Vision Research, 15, 1975. 161-171.
}
\author{Jose Gama}
\examples{
data(dichromaticCopunctalPoint)
dichromaticCopunctalPoint
}
\keyword{datasets}

