\name{example1Lanthony1978}
\alias{example1Lanthony1978}
\title{Example of cap arrangements for the D-15d test, Simple/Extreme Anomalous Trichromacy}
\usage{example1Lanthony1978}
\description{ \code{example1Lanthony1978} contains an example of cap arrangements for the D-15d test, Simple/Extreme Anomalous Trichromacy, from Lanthony (1978)}
\format{
  This data frame contains the following columns:
  \describe{
    \item{SimpleAnomalousTrichromacyD15}{ example cap arrangements D15 - Simple Anomalous Trichromacy}
    \item{SimpleAnomalousTrichromacyD15d}{ example cap arrangements D15d - Simple Anomalous Trichromacy}
    \item{ExtremeAnomalousTrichromacyD15}{ example cap arrangements D15 - Extreme Anomalous Trichromacy}
    \item{ExtremeAnomalousTrichromacyD15d}{ example cap arrangements D15d - Extreme Anomalous Trichromacy}
  }
}
\source{
The Desaturated Panel D-15 P. Lanthony Documenta Ophthalmologica 46,1: 185-189, 1978
}
\references{
The Desaturated Panel D-15 P. Lanthony Documenta Ophthalmologica 46,1: 185-189, 1978
}
\author{Jose Gama}
\examples{
data(example1Lanthony1978)
example1Lanthony1978
}
\keyword{datasets}

