\name{example2Lanthony1978}
\alias{example2Lanthony1978}
\title{Example of cap arrangements for the D-15d test, Central Serous Choroidopathy/Optic Neuritis/Autosomal Dominant OpticAtrophy}
\usage{example2Lanthony1978}
\description{ \code{example2Lanthony1978} contains an example of cap arrangements for the D-15d test, Central Serous Choroidopathy/Optic Neuritis/Autosomal Dominant OpticAtrophy, from Lanthony (1978)}
\format{
  This data frame contains the following columns:
  \describe{
    \item{CentralSerousChoroidopathyD15}{ example cap arrangements D15 - Central Serous Choroidopathy}
    \item{CentralSerousChoroidopathyD15d}{ example cap arrangements D15d - Central Serous Choroidopathy}
    \item{OpticNeuritisD15}{ example cap arrangements D15 - Optic Neuritis}
    \item{OpticNeuritisD15d}{ example cap arrangements D15d - Optic Neuritis}
    \item{AutosomalDominantOpticAtrophyD15}{ example cap arrangements D15 - Autosomal Dominant OpticAtrophy}
    \item{AutosomalDominantOpticAtrophyD15d}{ example cap arrangements D15d - Autosomal Dominant OpticAtrophy}
  }
}
\source{
THE DESATURATED PANEL D-15 P. LANTHONY Documenta Ophthalmologica 46,1: 185-189, 1978
}
\references{
THE DESATURATED PANEL D-15 P. LANTHONY Documenta Ophthalmologica 46,1: 185-189, 1978
}
\author{Jose Gama}
\examples{
data(example2Lanthony1978)
example2Lanthony1978
}
\keyword{datasets}
