\name{exampleBowman1982}
\alias{exampleBowman1982}
\title{Example of cap arrangements for the D-15d test}
\usage{exampleBowman1982}
\description{ \code{exampleBowman1982} contains an example of cap arrangements for the D-15d test, from Bowman (1982)}
\format{
  This data frame contains the following columns:
  \describe{
    \item{A}{ example cap arrangements A}
    \item{B}{ example cap arrangements B}
    \item{C}{ example cap arrangements C}
    \item{D}{ example cap arrangements D}
    \item{E}{ example cap arrangements E}
    \item{F}{ example cap arrangements F}
  }
}
\source{
A Method For Quantitative Scoring Of The Farnsworth Panel D-15 K.J. Bowman 1982
}
\references{
A Method For Quantitative Scoring Of The Farnsworth Panel D-15 K.J. Bowman 1982
}
\author{Jose Gama}
\examples{
data(exampleBowman1982)
exampleBowman1982
}
\keyword{datasets}
