\name{exampleFarnsworth1974}
\alias{exampleFarnsworth1974}
\title{Example of cap arrangements for the D-15 test, deuteranope/protanope/tritanope}
\usage{exampleFarnsworth1974}
\description{ \code{exampleFarnsworth1974} contains an example of cap arrangements for the D-15 test, deuteranope/protanope/tritanope, from Farnsworth (1974)}
\format{
  This data frame contains the following columns:
  \describe{
    \item{deuteranope}{ example cap arrangements D15 - deuteranope}
    \item{protanope}{ example cap arrangements D15 - protanope}
    \item{tritanope}{ example cap arrangements D15 - tritanope}
  }
}
\source{
Farnsworth D. The Farnsworth Dichotomous Test for Color Blindness. Panel D-15. New York, Psychological Testing, 1974
}
\references{
Farnsworth D. The Farnsworth Dichotomous Test for Color Blindness. Panel D-15. New York, Psychological Testing, 1974
}
\author{Jose Gama}
\examples{
data(exampleFarnsworth1974)
exampleFarnsworth1974
}
\keyword{datasets}
