\name{exampleFM100}
\alias{exampleFM100}
\title{Example of cap arrangements for the FM-100 test}
\usage{exampleFM100}
\description{ \code{exampleFM100} contains an example of cap arrangements for the FM-100, from Hidayat (2008)}
\format{
  This table contains one example of cap arrangements for the FM-100
}
\source{
proceedings of the New Zealand Generating fast automated reports for the Farnsworth-Munsell 100-hue colour vision test
Ray Hidayat, Computer Science Research Student Conference 2008
}
\references{
proceedings of the New Zealand Generating fast automated reports for the Farnsworth-Munsell 100-hue colour vision test
Ray Hidayat, Computer Science Research Student Conference 2008
}
\author{Jose Gama}
\examples{
data(exampleFM100)
exampleFM100
}
\keyword{datasets}
