\name{exampleNRC1981}
\alias{exampleNRC1981}
\title{Example of cap arrangements for the D-15 test, protanope/deuteranope/monochromat}
\usage{exampleNRC1981}
\description{ \code{exampleNRC1981} contains an example of cap arrangements for the D-15d test, protanope/deuteranope/monochromat, from National Research Council (1981)}
\format{
  This data frame contains the following columns:
  \describe{
    \item{protanope}{ example cap arrangements D15 - protanope}
    \item{deuteranope}{ example cap arrangements D15 - deuteranope}
    \item{monochromat}{ example cap arrangements D15 - monochromat}
  }
}
\source{
Procedures for Testing Color Vision: Report of Working Group 41, 1981, Committee on Vision, National Research Council, pp. 107
}
\references{
Procedures for Testing Color Vision: Report of Working Group 41, 1981, Committee on Vision, National Research Council, pp. 107
}
\author{Jose Gama}
\examples{
data(exampleNRC1981)
exampleNRC1981
}
\keyword{datasets}
