\name{exampleSimunovic2004}
\alias{exampleSimunovic2004}
\title{Example of cap arrangements for the D-15 test, rodMonochromat/blueConeMonochromat}
\usage{exampleSimunovic2004}
\description{ \code{exampleSimunovic2004} contains an example of cap arrangements for the D-15d test, rodMonochromat/blueConeMonochromat, from Lanthony (1978)}
\format{
  This data frame contains the following columns:
  \describe{
    \item{rodMonochromat}{ example cap arrangements D15 - rodMonochromat}
    \item{blueConeMonochromat}{ example cap arrangements D15 - blueConeMonochromat}
  }
}
\source{
Cone dystrophies Part 2 Cone dysfunction syndromes, Matthew P Simunovic
}
\references{
Cone dystrophies Part 2 Cone dysfunction syndromes, Matthew P Simunovic
}
\author{Jose Gama}
\examples{
data(exampleSimunovic2004)
exampleSimunovic2004
}
\keyword{datasets}
