\name{FarnsworthD15}
\alias{FarnsworthD15}
\title{Farnsworth D-15 cap colors}
\usage{FarnsworthD15}
\description{\code{FarnsworthD15} contains the cap colors for the D-15 tests, in CIELab and RGB from Farnsworth D (1947)
The Farnsworth Dichotomous test (D-15) is a short test for detecting congenital color vision deficiencies.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{CapNo}{ Cap Number}
    \item{Munsell}{ Munsell color}
    \item{x.C}{ CIE x cap color}
    \item{y.C}{ CIE y cap color}
    \item{R}{ R channel cap color}
    \item{G}{ G channel cap color}
    \item{B}{ B channel cap color}
 }
}
\source{
Farnsworth D. The Farnsworth Dichotomous Test for Color Blindness Panel D-15 Manual. New York, The Psychological Corp., 1947, pp. 1-8.
}
\references{
Farnsworth D. The Farnsworth Dichotomous Test for Color Blindness Panel D-15 Manual. New York, The Psychological Corp., 1947, pp. 1-8.
}
\author{Jose Gama}
\examples{
data(FarnsworthD15)
FarnsworthD15
}
\keyword{datasets}

