\name{H16}
\alias{H16}
\title{Farnsworth H-16 cap colors}
\usage{H16}
\description{\code{H16} contains the cap colors for the Farnsworth H-16 test, in Yxy coordinates.
The Farnsworth H-16 is a short test for detecting congenital color vision deficiencies.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{CapNo}{ Cap Number}
    \item{x.C}{ CIE x cap color}
    \item{y.C}{ CIE y cap color}
    \item{Munsell}{ Munsell color}
    \item{ProductionNo}{ Munsell Production Number}
 }
}
\source{
Judd, D.B. and MacAdam, D.L., 1979
Contributions to Color Science
University of Rochester. Institute of Optics and Center for Building Technology
Department of Commerce, National Bureau of Standards
}
\references{
Judd, D.B. and MacAdam, D.L., 1979
Contributions to Color Science
University of Rochester. Institute of Optics and Center for Building Technology
Department of Commerce, National Bureau of Standards
}
\author{Jose Gama}
\examples{
data(H16)
H16
}
\keyword{datasets}
