\name{LanthonyD15}
\alias{LanthonyD15}
\title{Farnsworth H-16 cap colors}
\usage{LanthonyD15}
\description{\code{LanthonyD15} contains the cap colors for Lanthony D-15 test, in Yxy coordinates.
The Lanthony D-15 (desaturated D-15) is a short test for detecting congenital color vision deficiencies.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{CapNo}{ Cap Number}
    \item{Munsell}{ Munsell color}
    \item{x.C}{ CIE x cap color}
    \item{y.C}{ CIE y cap color}
    \item{R}{ R channel cap color}
    \item{G}{ G channel cap color}
    \item{B}{ B channel cap color}
 }
}
\source{
Judd, D.B. and MacAdam, D.L., 1979
Contributions to Color Science
University of Rochester. Institute of Optics and Center for Building Technology
Department of Commerce, National Bureau of Standards
}
\references{
Judd, D.B. and MacAdam, D.L., 1979
Contributions to Color Science
University of Rochester. Institute of Optics and Center for Building Technology
Department of Commerce, National Bureau of Standards
}
\author{Jose Gama}
\examples{
data(LanthonyD15)
LanthonyD15
}
\keyword{datasets}
