\name{neutralPoint}
\alias{neutralPoint}
\title{Neutral points for CIE 1976 uv, CIE 1931 xy and CIE 1960 uv}
\usage{neutralPoint}
\description{ \code{neutralPoint} contains the neutral points for CIE 1976 uv, CIE 1931 xy and CIE 1960 uv}
\format{
  This data frame contains the following columns:
  \describe{
    \item{CIE1931xy}{neutral point CIE 1931 xy}
    \item{CIE1960uv}{neutral point CIE 1976 uv}
    \item{CIE1960uv}{neutral point CIE 1960 uv}
  }
}
\author{Jose Gama}
\examples{
data(neutralPoint)
neutralPoint
}
\keyword{datasets}

