\name{plotConfusionVectors}
\alias{plotConfusionVectors}
\title{Plot confusion vectors for CIE 1976 uv, CIE 1931 xy and CIE 1960 uv}
\description{\code{plotConfusionVectors} Plots the confusion vectors for
 CIE 1976 uv, CIE 1931 xy and CIE 1960 uv.
}
\usage{
plotConfusionVectors(colorSpace='CIE1931xy')
}
\arguments{
  \item{colorSpace}{ chosen colorSpace, default='CIE1931xy'}
}
\value{
  none
}
\author{Jose Gama}
\examples{
# find duplicate values
\dontrun{plotConfusionVectors()
}
}
\keyword{programming}
