\name{Roth28}
\alias{Roth28}
\title{Roth-28 cap colors}
\usage{Roth28}
\description{\code{Roth28} contains the cap colors for the Roth-28 tests, in CIELab and RGB from Roth A (1966)
The Roth-28 is a short test for detecting congenital color vision deficiencies.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{CIEL}{ CIELab L channel cap color}
    \item{CIEa}{ CIELab a channel cap color}
    \item{CIEb}{ CIELab b channel cap color}
    \item{R}{ R channel cap color}
    \item{G}{ G channel cap color}
    \item{B}{ B channel cap color}
 }
}
\source{
Roth A. Test-28 hue de Roth selon Farnsworth–Munsell (Manual). Paris: Luneau, 1966.
}
\references{
Roth A. Test-28 hue de Roth selon Farnsworth–Munsell (Manual). Paris: Luneau, 1966.
}
\author{Jose Gama}
\examples{
data(Roth28)
Roth28
}
\keyword{datasets}

