\name{typicalD15}
\alias{typicalD15}
\title{Typical cap arrangements for the D-15 tests}
\usage{typicalD15}
\description{ \code{typicalD15} contains typical cap arrangements for the D-15 tests, from Farnsworth D (1947), Simunovic (1998) and NRC (1981)}
\format{
  This data frame contains the following columns:
  \describe{
    \item{protanope}{ typical cap arrangements - protanope}
    \item{deuteranope}{ typical cap arrangements - deuteranope}
    \item{tritanope}{ typical cap arrangements - tritanope}
    \item{monochromat}{ typical cap arrangements - monochromat}
    \item{rodMonochromat}{ typical cap arrangements - rodMonochromat}
    \item{blueConeMonochromat}{ typical cap arrangements - blueConeMonochromat}
  }
}
\source{
Farnsworth D. The Farnsworth Dichotomous Test for Color Blindness Panel D-15 Manual. New York, The Psychological Corp., 1947, pp. 1-8.
Simunovic MP, Moore AT. The cone dystrophies. Eye 1998;12:553–65.
National Research Council (US). Committee on Vision. Procedures for testing color vision: report of Working Group 41. National Academies Press, 1981.
}
\references{
Farnsworth D. The Farnsworth Dichotomous Test for Color Blindness Panel D-15 Manual. New York, The Psychological Corp., 1947, pp. 1-8.
Simunovic MP, Moore AT. The cone dystrophies. Eye 1998;12:553–65.
National Research Council (US). Committee on Vision. Procedures for testing color vision: report of Working Group 41. National Academies Press, 1981.
}
\author{Jose Gama}
\examples{
data(typicalD15)
typicalD15
}
\keyword{datasets}

