\name{VKStable2}
\alias{VKStable2}
\title{Table with results of D-15 tests scored with the Vingrys and King-Smith method}
\usage{VKStable2}
\description{ \code{VKStable2} contains tthe results of D-15 tests scored with the Vingrys and King-Smith method, from Vingrys and King-Smith (1988), table 2}
\format{
  This data frame contains the following columns:
  \describe{
    \item{typeCVD}{ Type of color vision}
    \item{sample}{ Number in sample}
    \item{Angle}{ Angle}
    \item{S.index}{ S-index}
    \item{C.index}{ C-index}
    \item{Major}{ Major radius}
    \item{Minor}{ Minor radius}
    \item{TES}{ TES}
    \item{TCDS}{ TCDS}
  }
}
\source{
Atchison DA, Bowman KJ, Vingrys AJ
Quantitave scoring methods for D15 panel tests in the diagnosis of congenital colour-vision deficiencies.
Optometry and Vision Science 1991, 68:41-48.
}
\references{
Atchison DA, Bowman KJ, Vingrys AJ
Quantitave scoring methods for D15 panel tests in the diagnosis of congenital colour-vision deficiencies.
Optometry and Vision Science 1991, 68:41-48.
}
\author{Jose Gama}
\examples{
data(VKStable2)
VKStable2
}
\keyword{datasets}

