\name{FarnsworthMunsell100Hue}
\alias{FarnsworthMunsell100Hue}
\title{Farnsworth D-15 cap colors}
\usage{FarnsworthMunsell100Hue}
\description{\code{FarnsworthMunsell100Hue} contains the cap colors for the Farnsworth Munsell 100-Hue tests, in CIELab and RGB from Farnsworth D (1957)
The Farnsworth Munsell 100-Hue is a test for detecting congenital and acquired color vision deficiencies.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{CIEL}{ CIELab L channel cap color}
    \item{CIEa}{ CIELab a channel cap color}
    \item{CIEb}{ CIELab b channel cap color}
    \item{R}{ R channel cap color}
    \item{G}{ G channel cap color}
    \item{B}{ B channel cap color}
 }
}
\source{
Farnsworth D: The Farnsworth-Munsell 100-Hue Test for the Examination of Color Discrimination Manual. Baltimore, Munsell Color Co., 1957, pp. 1-7.
}
\references{
Farnsworth D: The Farnsworth-Munsell 100-Hue Test for the Examination of Color Discrimination Manual. Baltimore, Munsell Color Co., 1957, pp. 1-7.
}
\author{Jose Gama}
\examples{
data(FarnsworthMunsell100Hue)
FarnsworthMunsell100Hue
}
\keyword{datasets}

